Announce Discontinuation
As sad as I may be, but this project has been left barren for months now, and no one seems to be willing to add to this. I may pick this up in the future, I may not. But this is to be left untouched for the time being, serving as a time capsule representing the memories and moments its devs created along with the project. ;)